4th round of standard compiler fixes

+++ src/d/actor/d_a_e_kr.cpp
@@ -2233,7 +2233,7 @@ static int useHeapInit(fopAc_ac_c* i_this) {
     }
 
     J3DModel* model = kr->mpMorf->getModel();
-    model->setUserArea((int)kr);
+    model->setUserArea((intptr_t)kr);
 
     for (u16 i = 0; i < model->getModelData()->getJointNum(); i++) {
         if (i != 0) {
