+++ src/d/actor/d_a_e_yr.cpp
@@ -2407,7 +2407,7 @@ static int useHeapInit(fopAc_ac_c* i_this) {
     }
 
     J3DModel* model = yr->mpMorfSO->getModel();
-    model->setUserArea((int)yr);
+    model->setUserArea((intptr_t)yr);
 
     for (u16 i = 0; i < model->getModelData()->getJointNum(); i++) {
         if (i != 0) {
